Finished: Organize the components into utility ones and domain ones.
