+++ src/lib/UI/domain-components/cards/CoffeeBeansCard.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import type { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
-  import FlexRow from "$lib/UI/FlexRow.svelte";
-  import Card from "./Card.svelte";
+  import Card from "$lib/UI/utility-components/Card.svelte";
+  import FlexRow from "$lib/UI/utility-components/FlexRow.svelte";
 
   // Props:
