+++ src/lib/UI/utility-components/modals/CloseButton.svelte
