+++ src/routes/coffee_beans/[name]/EditCoffeeBeansModal.svelte
@@ -7,10 +7,10 @@
   import { editCoffeeBeans } from "$lib/database/v1/indexedDB";
   import { CoffeeBeans, CoffeeBeansEditSubmit } from "$lib/domain/entities/CoffeeBeans";
   import { routes } from "$lib/domain/routes";
-  import ModalHeader from "$lib/UI/CoffeeBeansSelect/components/ModalHeader.svelte";
-  import Label from "$lib/UI/forms/Label.svelte";
-  import Textarea from "$lib/UI/forms/Textarea.svelte";
   import MySidebar from "$lib/UI/layout/components/MySidebar.svelte";
+  import Label from "$lib/UI/utility-components/forms/Label.svelte";
+  import Textarea from "$lib/UI/utility-components/forms/Textarea.svelte";
+  import ModalHeader from "$lib/UI/utility-components/modals/ModalHeader.svelte";
 
   // Props:
