+++ src/routes/recipes/+page.svelte
@@ -4,10 +4,10 @@
   import type { Recipe } from "$lib/domain/entities/Recipe";
   import { sortRecipesByTimestampDesc as byTimestampDesc } from "$lib/domain/helpers/sortRecipes";
   import { routes } from "$lib/domain/routes";
-  import RecipeCard from "$lib/UI/cards/RecipeCard.svelte";
-  import AddDemoCoffeeBeans_PageBlock from "$lib/UI/domain-components/AddDemoCoffeeBeans_PageBlock.svelte";
+  import RecipeCard from "$lib/UI/domain-components/cards/RecipeCard.svelte";
+  import MyFab from "$lib/UI/domain-components/FABs/AddRecipeFab.svelte";
+  import AddDemoCoffeeBeans_PageBlock from "$lib/UI/domain-components/page-blocks/AddDemoCoffeeBeans_PageBlock.svelte";
   import PageHeadline from "$lib/UI/layout/PageHeadline.svelte";
-  import MyFab from "$lib/UI/MyFab.svelte";
 
   // State:
