+++ src/routes/recipes/add/CoffeeBeansSelect.svelte
@@ -4,8 +4,8 @@
 
 <script lang="ts">
   import { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
-  import NewCoffeeBeansModal from "$lib/UI/CoffeeBeansSelect/NewCoffeeBeansModal.svelte";
-  import Label from "$lib/UI/forms/Label.svelte";
+  import NewCoffeeBeansModal from "$lib/UI/domain-components/modals/NewCoffeeBeansModal.svelte";
+  import Label from "$lib/UI/utility-components/forms/Label.svelte";
 
   // Props:
